+++ scripts/exporters/cards-exporter.mjs
@@ -20,9 +20,12 @@ export class CardsExporter extends AbstractExporter {
     const documents = await this.pack.getIndex();
 
     for (const indexDocument of documents) {
-      this.dataset.entries[indexDocument.name] = CardsExporter.getDocumentData(
-        indexDocument,
-        await this.pack.getDocument(indexDocument._id),
+      this.dataset.entries[indexDocument.name] = foundry.utils.mergeObject(
+        CardsExporter.getDocumentData(
+          indexDocument,
+          await this.pack.getDocument(indexDocument._id),
+        ),
+        this.existingContent[indexDocument.name],
       );
 
       this._stepProgressBar();