+++ scripts/exporters/rolltable-exporter.mjs
@@ -20,9 +20,12 @@ export class RollTableExporter extends AbstractExporter {
     const documents = await this.pack.getIndex();
 
     for (const indexDocument of documents) {
-      this.dataset.entries[indexDocument.name] = RollTableExporter.getDocumentData(
-        indexDocument,
-        await this.pack.getDocument(indexDocument._id),
+      this.dataset.entries[indexDocument.name] = foundry.utils.mergeObject(
+        RollTableExporter.getDocumentData(
+          indexDocument,
+          await this.pack.getDocument(indexDocument._id),
+        ),
+        this.existingContent[indexDocument.name] ?? {},
       );
 
       this._stepProgressBar();